Support saving found blocks and round effort to preserve stats across runs
If `statsDir` is set in the config file, create a file `stats.json` and
periodically update it with stats. When the program is loaded the file
will be read if it exists.
