+++ .gitignore
@@ -20,3 +20,4 @@ mainnet.json
 
 /build/_workspace/
 /build/bin/
+.idea
