+++ README.md
@@ -146,11 +146,18 @@ Configuration is self-describing, just copy *config.example.json* to *config.jso
       "port": 18081,
       "timeout": "10s"
     }
-  ]
+  ],
+  
+  "statsDir": "stats"
 }
 ```
 
 You must use `anything.WorkerID` as username in your miner. Either disable address validation or use `<address>.WorkerID` as username. If there is no workerID specified your rig stats will be merged under `0` worker. If mining software contains dev fee rounds its stats will usually appear under `0` worker. This stratum acts like your own pool, the only exception is that you will get rewarded only after block found, shares only used for stats.
+
+## Stats
+Blocks found and round progress will be saved across runs in `$statsDir/stats.json` if you have enabled them by setting 
+`statsDir` in your `config.json`. You can also choose the interval (seconds) that stats are collected with by setting 
+`statsInterval`. The default is 5 seconds. 
 
 ### Donations
